+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.0.0-dev.6
+- Add defaultRoute
+
+## 1.0.0-dev.5
+- Add @optional new initialRoute to PlusRouteInformationParser
+- Add @optional custom loading page to PlusRouterDelegate
+- Fix error when reload page
+
+## 1.0.0-dev.4
+- Update README.md
+
 ## 1.0.0-dev.3
 - Add navigate back()
 - Add navigate canBack()